+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Add missing inherit of `grim.annotations.Annotations` gwt module into `org.realityforge.vecmath.Vecmath`.
+* Add a `Vector3f.angle(Vector3f)` method to calculate the angle between two vectors.
 
 ### [v0.01](https://github.com/realityforge/vecmath/tree/v0.01) (2020-10-03)
 [Full Changelog](https://github.com/realityforge/vecmath/compare/c629a14f7295a371410e3a500d252917a8ee0c33...v0.01)